requirements.txt
azure-messaging-webpubsubservice==1.0.0
azure.identity
websockets
azure-cognitiveservices-speech
urllib3
flask
python-dotenv
pip-system-certs
openai
flask-socketio
pydub